// src/shared/types/domain.ts
export enum Rank {
    White = 'White',
    Yellow = 'Yellow',
    Orange = 'Orange',
    Green = 'Green',
    Blue = 'Blue',
    Brown = 'Brown',
    Dan1 = 'Black (DAN 1)',
    Dan2 = 'Black (DAN 2)',
    Dan3 = 'Black (DAN 3)',
    Dan4 = 'Black (DAN 4)',
    Dan5 = 'Black (DAN 5)',
    Dan6 = 'Black (DAN 6)',
    Dan7 = 'Black (DAN 7)',
}

export enum ActivityStatus {
    Constant = 'Constant',
    Intermittent = 'Intermittent',
    Dormant = 'Dormant',
}